% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{aveThetas2}
\alias{aveThetas2}
\title{Computes the AVE, DWAVE, WAVE, CH-EXP and CH-ECDF methods for multivariate GLMMs (m-GLMMs)}
\usage{
aveThetas2(Models, ModelsOne, Data, GHk = 5, n, Q, extraParam, extraParamOne,
  m = 1)
}
\arguments{
\item{Models}{A list which contains the lme4 model objects taken from the pairwise separate estimations (list of size Q*(Q-1)/2)}

\item{ModelsOne}{A list which contains the lme4 model objects taken from the univariate separate estimations (list of size Q)}

\item{Data}{a data.frame with the data. 1st column id, 2nd column time, remaining Q columns are the y 0/1 values (Q items)}

\item{GHk}{Number of Gauss-Hermite points per dimension of integration}

\item{n}{number of individuals}

\item{Q}{the number of items. Set this to four.}

\item{extraParam}{a helper list which depends on pairwise estimates}

\item{extraParamOne}{a helper list which depends on univariate estimates}

\item{m}{an integer which is useful for subsequent runs}
}
\value{
The estimated parameters of the model with methods AVE, DWAVE, WAVE, CH-EXP and CH-ECDF
}
\description{
It acts on the Models structure for all pairs of items and returns the estimates for the m-GLMM parameters with 5 different methods
}

