% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoExample2.R
\name{demoExample2}
\alias{demoExample2}
\title{A Demo Example of the CHmGLMM package}
\usage{
demoExample2()
}
\value{
Returns the demo example output. It runs as a script with no input.
}
\description{
It sets up an estimation problem, and performs the estimation with methods AVE, DWAVE, WAVE, CH-EXP and CH-ECDF
}

