% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateModelFit2.R
\name{estimateModelFit2}
\alias{estimateModelFit2}
\title{Estimates the modelFit}
\usage{
estimateModelFit2(Data, Q, n)
}
\arguments{
\item{Data}{a data.frame with the data. 1st column id, 2nd column time, remaining Q columns are the y 0/1 values (Q items)}

\item{Q}{the number of items. Set this to four.}

\item{n}{number of individuals}
}
\value{
Returns the model fitted
}
\description{
It acts on the Data data.frame, on the number of items Q and on the number of individuals
}

