% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData2.R
\name{generateData2}
\alias{generateData2}
\title{Generates the Data}
\usage{
generateData2(id, times, n, X, Z, betas, b, Q)
}
\arguments{
\item{id}{individual number}

\item{times}{one-dimensional array of time points}

\item{n}{number of individuals}

\item{X}{fixed effects design matrix. Usually includes an intercept and time.}

\item{Z}{random effects design matrix. Usually includes an intercept and time}

\item{betas}{the fixed effects parameters for the intercept and the slope.}

\item{b}{the random effects taken from a multivariate normal with mean 0 and variance matrix D}

\item{Q}{the number of items. Set this to four.}
}
\value{
Returns the generated Data as data.frame
}
\description{
It acts on the primitives of the estimation problem and returns the Data
}

